+++ CHANGELOG.md
@@ -4,6 +4,30 @@
 ### Breaking changes
 - From `Projection.Model` removed the validation mechanism provided by the [`ember-validations`](https://github.com/DockYard/ember-validations) addon.
 
+## [0.2.0-beta.2] - 2018-10-17
+### Fixed
+- `Serializer.OData`:
+    - The `normalize` method for update type in case of hasMany polymorphic relationships.
+
+## [0.2.0-beta.1] - 2018-10-10
+### Added
+- Query language:
+    - `GeometryPredicate` for geometry data types.
+
+## [0.2.0-beta.0] - 2018-09-25
+### Added
+- `Serializer.Base`:
+    - The `normalize` method for update typeClass in case of polymorphic relationships.
+
+## [0.13.0-beta.0] - 2018-09-17
+### Added
+- Query language:
+    - The `IsOfPredicate` with implementation for `OData`.
+    - The `isOf` method in `Query.Builder` for simple usage `IsOfPredicate`.
+
+### Fixed
+- Loss of `intersectsValue` in the `GeographyPredicate` when it inside `NotPredicate`.
+
 ## [0.12.0] - 2018-05-28
 ### Added
 - `Projection.OnlineStore` as online store for `Offline.Store`.